Crate keys

TiKV key building

Modules

rewrite
Key rewriting

Enums

Error

Constants

APPLY_STATE_SUFFIX

DATA_MAX_KEY

DATA_MIN_KEY

DATA_PREFIX

DATA_PREFIX_KEY

EMPTY_KEY

LOCAL_MAX_KEY

LOCAL_MIN_KEY

LOCAL_PREFIX

MAX_KEY

MIN_KEY

PREPARE_BOOTSTRAP_KEY

RAFT_LOG_SUFFIX

RAFT_STATE_SUFFIX

REGION_META_MAX_KEY

REGION_META_MIN_KEY

REGION_META_PREFIX

REGION_META_PREFIX_KEY

REGION_RAFT_MAX_KEY

REGION_RAFT_MIN_KEY

REGION_RAFT_PREFIX

REGION_RAFT_PREFIX_KEY

REGION_STATE_SUFFIX

SNAPSHOT_RAFT_STATE_SUFFIX

STORE_IDENT_KEY

Functions

apply_state_key

data_end_key

data_key

decode_raft_key
Get the region id and log type from raft key.

decode_raft_log_key
Get the region id and index from raft log key generated by raft_log_key.

decode_region_key
Decode region key, return the region id and meta suffix type.

decode_region_meta_key
Decode region meta key, return the region id and meta suffix type.

decode_region_raft_key
Decode region raft key, return the region id and raft suffix type.

enc_end_key
Get the end_key of current region in encoded form.

enc_start_key
Get the start_key of current region in encoded form.

make_region_key

make_region_meta_key

make_region_prefix

next_key
Computes the next key of the given key.

next_key_no_alloc

origin_end_key

origin_key

raft_log_index
Get the log index from raft log key generated by raft_log_key.

raft_log_key

raft_log_prefix

raft_state_key

region_meta_prefix

region_raft_prefix

region_raft_prefix_len

region_state_key

snapshot_raft_state_key

validate_data_key

Type Definitions

Result