Struct keys::Private

pub struct Private {
    pub network: Network,
    pub secret: Secret,
    pub compressed: bool,
}

Secret with additional network identifier and format type

Fields
network: Network

The network on which this key should be used.

secret: Secret

ECDSA key.

compressed: bool

True if this private key represents a compressed address.

Methods

impl Private

pub fn sign(&self, message: &Message) -> Result<Signature, Error>

pub fn sign_compact(&self, message: &Message) -> Result<CompactSignature, Error>